// careerladder_web/career_ladder/types/performance.ts
export type Performance = {
    comments: string
    employer_id: string
    company_name: string
    companyLogo: string
    email: string
    communication: number
    overall_rating: number
    problem_solving: number
    professionalism: number
    teamwork: number
    technical_skills: number
}